start detail sreen

+++ composeApp/src/commonMain/kotlin/com/rykova_e/kts_project/data/source/remote/model/ReviewRemote.kt
@@ -12,7 +12,7 @@ data class ReviewRemote(
     @SerialName("average")
     val averageReview: String,
     @SerialName("count")
-    val count: Long,
+    val count: Int,
     @SerialName("distribution")
     val distribution: List<Int>
 )
